Refactor #575: federation / following 系を handler_stubs から分離
federation・following package の handler_stubs.go / handler_stubs_test.go
を完全廃止し、テーマ別ファイルに再編する。中身は既に本実装済みで、
ファイル名のみが misleading な "stub" だったため、#574 (PR #581) で
admin に対して実施したのと同じ disassembly を適用する。

federation:
- host_listings.go    — Followers / Following / Users + parseHostPage helpers
- stats.go            — Stats + totalCounts
- update_remote_user.go — UpdateRemoteUser
- 同名 _test.go に対応分割。postBody / postStub helper を handler_test.go に集約

following:
- relation.go         — Invalidate / UpdateFollow / UpdateFollowAll
- requests_sent.go    — RequestsSent
- 同名 _test.go に対応分割。重複していた postStub helper は handler_test.go の
  既存 postJSON に統合

Coverage:
- internal/api/federation: 92.2%
- internal/api/following:  95.0%

Closes #575
