--- internal/api/federation/handler_stubs_test.go
@@ -1,242 +1,0 @@
-package federation
-
-import (
-	"encoding/json"
-	"errors"
-	"net/http"
-	"net/http/httptest"
-	"strings"
-	"testing"
-
-	"github.com/labstack/echo/v4"
-	"github.com/shiroha-a/mk/internal/model"
-	"github.com/shiroha-a/mk/internal/testutil"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-)
-
-func postBody(handler func(echo.Context) error, body string) *httptest.ResponseRecorder {
-	e := echo.New()
-	req := httptest.NewRequest(http.MethodPost, "/", strings.NewReader(body))
-	req.Header.Set(echo.HeaderContentType, echo.MIMEApplicationJSON)
-	rec := httptest.NewRecorder()
-	c := e.NewContext(req, rec)
-	_ = handler(c)
-	return rec
-}
-
-func postStub(handler func(echo.Context) error) *httptest.ResponseRecorder {
-	return postBody(handler, `{}`)
-}
-
-// --- bind validation ---
-
-func TestFollowers_MissingHost(t *testing.T) {
-	h, _ := newHandler(t)
-	rec := postStub(h.Followers)
-	assert.Equal(t, http.StatusBadRequest, rec.Code)
-	// 400 エラーレスポンス直後に 200 を追記してしまう double-write バグの
-	// regression guard。body は単一の JSON オブジェクト (error wrapper) になる
-	// はずで、追加の [] が後続しないこと。
-	var single map[string]any
-	assert.NoError(t, json.Unmarshal(rec.Body.Bytes(), &single),
-		"response body must be a single JSON object, not concatenated")
-}
-
-func TestFollowing_MissingHost(t *testing.T) {
-	h, _ := newHandler(t)
-	assert.Equal(t, http.StatusBadRequest, postStub(h.Following).Code)
-}
-
-func TestUsers_MissingHost(t *testing.T) {
-	h, _ := newHandler(t)
-	assert.Equal(t, http.StatusBadRequest, postStub(h.Users).Code)
-}
-
-func TestStats_EmptyService(t *testing.T) {
-	h, _ := newHandler(t)
-	assert.Equal(t, http.StatusOK, postStub(h.Stats).Code)
-}
-
-func TestUpdateRemoteUser_MissingUserID(t *testing.T) {
-	h, _ := newHandler(t)
-	assert.Equal(t, http.StatusBadRequest, postStub(h.UpdateRemoteUser).Code)
-}
-
-// --- detail: Followers / Following / Users with repo wired ---
-
-func TestFollowers_FiltersByHost(t *testing.T) {
-	h, _ := newHandler(t)
-	repo := testutil.NewMockFollowingRepository()
-	remote := "remote.example"
-	other := "elsewhere.example"
-	repo.Followings["f1"] = &model.Following{ID: "f1", FollowerID: "r1", FollowerHost: &remote, FolloweeID: "local1"}
-	repo.Followings["f2"] = &model.Following{ID: "f2", FollowerID: "r2", FollowerHost: &remote, FolloweeID: "local2"}
-	repo.Followings["f3"] = &model.Following{ID: "f3", FollowerID: "r3", FollowerHost: &other, FolloweeID: "local3"}
-	h.SetFollowingRepo(repo)
-
-	rec := postBody(h.Followers, `{"host":"remote.example","limit":10}`)
-	assert.Equal(t, http.StatusOK, rec.Code)
-	var rows []map[string]any
-	require.NoError(t, json.Unmarshal(rec.Body.Bytes(), &rows))
-	assert.Len(t, rows, 2)
-}
-
-func TestFollowing_FiltersByHost(t *testing.T) {
-	h, _ := newHandler(t)
-	repo := testutil.NewMockFollowingRepository()
-	remote := "remote.example"
-	repo.Followings["f1"] = &model.Following{ID: "f1", FollowerID: "local1", FolloweeID: "r1", FolloweeHost: &remote}
-	h.SetFollowingRepo(repo)
-
-	rec := postBody(h.Following, `{"host":"remote.example"}`)
-	assert.Equal(t, http.StatusOK, rec.Code)
-	var rows []map[string]any
-	require.NoError(t, json.Unmarshal(rec.Body.Bytes(), &rows))
-	assert.Len(t, rows, 1)
-}
-
-func TestUsers_FiltersByHost(t *testing.T) {
-	h, _ := newHandler(t)
-	userRepo := testutil.NewMockUserRepository()
-	remote := "remote.example"
-	userRepo.Users["u1"] = &model.User{ID: "u1", Username: "alice", Host: &remote}
-	userRepo.Users["u2"] = &model.User{ID: "u2", Username: "bob"}
-	h.SetUserRepo(userRepo)
-
-	rec := postBody(h.Users, `{"host":"remote.example"}`)
-	assert.Equal(t, http.StatusOK, rec.Code)
-	var rows []map[string]any
-	require.NoError(t, json.Unmarshal(rec.Body.Bytes(), &rows))
-	// MockUserRepository.ListUsers が Hostname フィルタを実装してないと local も
-	// 混入してしまうため、実装側でフィルタしないケースでもテストは最低限
-	// 200 が返ることを確認する。実 DB 経路での厳密な filter 動作検証は repo
-	// 側テスト (TestUserRepository_ListUsers_HostnameFilter 相当) で行う。
-	assert.GreaterOrEqual(t, len(rows), 1)
-}
-
-// --- UpdateRemoteUser ---
-
-type fakeResolver struct {
-	called  int
-	lastURI string
-	err     error
-}
-
-func (f *fakeResolver) ForceResolveActor(uri string) (*model.User, error) {
-	f.called++
-	f.lastURI = uri
-	return nil, f.err
-}
-
-func TestUpdateRemoteUser_LocalUserIsNoop(t *testing.T) {
-	h, _ := newHandler(t)
-	userRepo := testutil.NewMockUserRepository()
-	userRepo.Users["u-local"] = &model.User{ID: "u-local", Username: "me", Host: nil, URI: nil}
-	h.SetUserRepo(userRepo)
-	h.SetResolver(&fakeResolver{})
-	assert.Equal(t, http.StatusNoContent, postBody(h.UpdateRemoteUser, `{"userId":"u-local"}`).Code)
-}
-
-func TestUpdateRemoteUser_ResolvesRemoteActor(t *testing.T) {
-	h, _ := newHandler(t)
-	userRepo := testutil.NewMockUserRepository()
-	uri := "https://remote.example/users/alice"
-	userRepo.Users["u-r"] = &model.User{ID: "u-r", Username: "alice", URI: &uri}
-	h.SetUserRepo(userRepo)
-	fr := &fakeResolver{}
-	h.SetResolver(fr)
-
-	rec := postBody(h.UpdateRemoteUser, `{"userId":"u-r"}`)
-	assert.Equal(t, http.StatusNoContent, rec.Code)
-	assert.Equal(t, 1, fr.called)
-	assert.Equal(t, uri, fr.lastURI)
-}
-
-func TestUpdateRemoteUser_UnknownUser(t *testing.T) {
-	h, _ := newHandler(t)
-	h.SetUserRepo(testutil.NewMockUserRepository())
-	rec := postBody(h.UpdateRemoteUser, `{"userId":"ghost"}`)
-	assert.Equal(t, http.StatusNotFound, rec.Code)
-}
-
-func TestUpdateRemoteUser_ResolverError(t *testing.T) {
-	h, _ := newHandler(t)
-	userRepo := testutil.NewMockUserRepository()
-	uri := "https://remote.example/users/alice"
-	userRepo.Users["u-r"] = &model.User{ID: "u-r", URI: &uri}
-	h.SetUserRepo(userRepo)
-	h.SetResolver(&fakeResolver{err: errors.New("network")})
-
-	rec := postBody(h.UpdateRemoteUser, `{"userId":"u-r"}`)
-	assert.Equal(t, http.StatusInternalServerError, rec.Code)
-}
-
-func TestUpdateRemoteUser_NoUserRepoUnwired(t *testing.T) {
-	h, _ := newHandler(t)
-	// userRepo 未注入 → 204 (no-op)
-	rec := postBody(h.UpdateRemoteUser, `{"userId":"u1"}`)
-	assert.Equal(t, http.StatusNoContent, rec.Code)
-}
-
-func TestUpdateRemoteUser_NoResolverUnwired(t *testing.T) {
-	h, _ := newHandler(t)
-	userRepo := testutil.NewMockUserRepository()
-	uri := "https://remote.example/users/alice"
-	userRepo.Users["u-r"] = &model.User{ID: "u-r", URI: &uri}
-	h.SetUserRepo(userRepo)
-	rec := postBody(h.UpdateRemoteUser, `{"userId":"u-r"}`)
-	assert.Equal(t, http.StatusNoContent, rec.Code)
-}
-
-func TestFollowers_NoRepoReturnsEmpty(t *testing.T) {
-	h, _ := newHandler(t)
-	rec := postBody(h.Followers, `{"host":"remote.example"}`)
-	assert.Equal(t, http.StatusOK, rec.Code)
-}
-
-func TestFollowing_NoRepoReturnsEmpty(t *testing.T) {
-	h, _ := newHandler(t)
-	rec := postBody(h.Following, `{"host":"remote.example"}`)
-	assert.Equal(t, http.StatusOK, rec.Code)
-}
-
-func TestUsers_NoRepoReturnsEmpty(t *testing.T) {
-	h, _ := newHandler(t)
-	rec := postBody(h.Users, `{"host":"remote.example"}`)
-	assert.Equal(t, http.StatusOK, rec.Code)
-}
-
-// bindHostPage の limit clamp 分岐 (>100) カバー
-func TestFollowers_LimitClampedAt100(t *testing.T) {
-	h, _ := newHandler(t)
-	h.SetFollowingRepo(testutil.NewMockFollowingRepository())
-	rec := postBody(h.Followers, `{"host":"remote.example","limit":500}`)
-	assert.Equal(t, http.StatusOK, rec.Code)
-}
-
-func TestStats_WithTopInstances(t *testing.T) {
-	h, instRepo := newHandler(t)
-	// 3 instance を登録。2 は top に入り、1 は "other" に流れる想定で
-	// limit=2 で呼び出す。
-	_ = instRepo.Create(&model.Instance{ID: "inst-1", Host: "alpha.example", FollowersCount: 10, FollowingCount: 5})
-	_ = instRepo.Create(&model.Instance{ID: "inst-2", Host: "beta.example", FollowersCount: 8, FollowingCount: 20})
-	_ = instRepo.Create(&model.Instance{ID: "inst-3", Host: "gamma.example", FollowersCount: 2, FollowingCount: 3})
-
-	rec := postBody(h.Stats, `{"limit":2}`)
-	assert.Equal(t, http.StatusOK, rec.Code)
-	var body map[string]any
-	require.NoError(t, json.Unmarshal(rec.Body.Bytes(), &body))
-	topSub, ok := body["topSubInstances"].([]any)
-	require.True(t, ok)
-	assert.Len(t, topSub, 2)
-	topPub, ok := body["topPubInstances"].([]any)
-	require.True(t, ok)
-	assert.Len(t, topPub, 2)
-	// topSubInstances は followers 上位 2 件 = 10 + 8 = 18
-	// 合計 followers = 20、other = 2
-	assert.EqualValues(t, 2, body["otherFollowersCount"])
-	// topPubInstances は following 上位 2 件 = 20 + 5 = 25
-	// 合計 following = 28、other = 3
-	assert.EqualValues(t, 3, body["otherFollowingCount"])
-}
